% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{excelToModelSST}
\alias{excelToModelSST}
\title{Load SST template}
\usage{
excelToModelSST(path = NULL, ...)
}
\arguments{
\item{path}{Either the path to the SST Template (standard SST calculation) or the vector of paths to the SST Templates (look-through calculation).}

\item{...}{Optional arguments passed to function.}
}
\value{
A model, of class sstCalculation.
}
\description{
Load the SST template(s) into a model. During this step, the content of the SST template is checked to ensure that the model is valid.

This function covers both the standard SST calculation (one SST-Template) and the case where the participation model is used (two or more SST-Templates).

For the look-through calculation, one needs to provide exactly one mother template and one (or multiple) daughter templates.
The order of the provided templates is irrelevant, as the mother and daughter templates are identified automatically.

Note that when the only provided template is a daughter template, this template will be considederd as a standard SST-Template (the additional daughter parameters are ignored).
}
