% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{sstCalculation-package}
\alias{sstCalculation-package}
\title{Implementation of the Swiss Solvency Test (SST) Standard Models.}
\description{
Framework for performing solvency related computations based on
 standard models for the Swiss Solvency Test (SST), a risk-based capital standard for Swiss
 insurance companies.
}
