% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{sstCalculation}
\alias{sstCalculation}
\title{Perform SST calculations}
\usage{
sstCalculation(path = NULL, nsim = NULL, seed = 1, openFDS = FALSE, ...)
}
\arguments{
\item{path}{Either the path to the SST Template (standard SST calculation) or the vector of paths to the SST Templates (look-through calculation). The mother and daughter are identified automatically.}

\item{nsim}{Number of simulations to perform.}

\item{seed}{Initial seed.}

\item{openFDS}{Either a logical value indicating if the resulting Excel template (FDS) should be opened right away in Excel, or a string containing the path where the template will be exported.}

\item{...}{Optional arguments passed to function.}
}
\value{
A list containing RBC simulations, market risk simulations, key figures.
}
\description{
Main function to perform SST calculations.

This function covers both the standard SST calculation (one SST-Template) and the case where the participation model is used (two or more SST-Templates).

For the look-through calculation, one needs to provide exactly one mother template and one (or multiple) daughter templates.
The order of the provided templates is irrelevant, as the mother and daughter templates are identified automatically.

Note that when the only provided template is a daughter template, this template will be considered as a standard SST-Template (the additional daughter parameters are ignored).
}
