% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{sstDashboard}
\alias{sstDashboard}
\title{Launching The Dashboard In A Browser}
\usage{
sstDashboard(workingDirectory = "~")
}
\arguments{
\item{workingDirectory}{Directory where the calculation results will be exported.}
}
\value{
None.
}
\description{
This function launches an interactive dashboard for SST computations.
}
